chore: Enable gosec and do some internal types modifications (#16)
Most of the errors from gosec stemmed from dangerous type conversions (e.g. from uint32 to uint16). Instead of doing checks everywhere inside the codebase on whether the type semantics are observed (e.g. checking that the uint32 is not larger than math.MaxUint16), I opted to handle this on the types definition and Babylon retrieval layer.

I made gosec exclude the testutil/ and itest/ directories as they are related to tests. Also, for the following gosec errors I put a flag to remove them from the linter:
- Two conversions from uint32 to uint16. Even though I performed the check previously, the static analyzer could not identify it.
- Go file inclusion. This file is not passed from the web, but the user's machine, so it doesn't need further checks.